requirements.txt
requests==2.11.1
python-whois-extended==0.6.10